Migration to separate component for recommendation actions
